pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hrules"
version = "0.1.0"
description = "HR-focused visibility compliance scanner: no hidden or low-visibility content"
authors = [{ name = "HRules Contributors" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
  "pillow>=10.0.0",
  "python-docx>=1.1.0",
  "pymupdf>=1.23.0",
  "pytesseract>=0.3.10",
  "beautifulsoup4>=4.12.2",
  "cssutils>=2.7.1",
  "psd-tools>=1.9.31",
  "reportlab>=4.0.4"
]

[project.scripts]
hrules = "hrules.cli:main"
hrules-gui = "hrules.gui:main"

[tool.setuptools.packages.find]
where = ["src"]
